# replace old libsecp256k1 fork with upstream & remove GMP dependency