--- a/README.md
+++ b/README.md
@@ -28,7 +28,6 @@ apt-get update && apt-get install   \
         jq                          \
         libboost-all-dev            \
         libcurl4-openssl-dev        \
-        libgmp-dev                  \
         libssl-dev                  \
         libtinfo5                   \
         libusb-1.0-0-dev            \